Check access rights

+++ src/main/kotlin/no/nav/klage/oppgave/api/controller/KlagebehandlingMeldingController.kt
@@ -9,6 +9,7 @@ import no.nav.klage.oppgave.api.view.MeldingModified
 import no.nav.klage.oppgave.api.view.MeldingView
 import no.nav.klage.oppgave.config.SecurityConfiguration.Companion.ISSUER_AAD
 import no.nav.klage.oppgave.repositories.InnloggetSaksbehandlerRepository
+import no.nav.klage.oppgave.service.KlagebehandlingService
 import no.nav.klage.oppgave.service.MeldingService
 import no.nav.klage.oppgave.util.getLogger
 import no.nav.klage.oppgave.util.logKlagebehandlingMethodDetails
@@ -25,6 +26,7 @@ class KlagebehandlingMeldingController(
     private val innloggetSaksbehandlerRepository: InnloggetSaksbehandlerRepository,
     private val meldingService: MeldingService,
     private val meldingMapper: MeldingMapper,
+    private val klagebehandlingService: KlagebehandlingService
 ) {
 
     companion object {
@@ -44,6 +46,8 @@ class KlagebehandlingMeldingController(
     ): MeldingCreated {
         val innloggetIdent = innloggetSaksbehandlerRepository.getInnloggetIdent()
         logKlagebehandlingMethodDetails("addMelding", innloggetIdent, klagebehandlingId, logger)
+
+        validateAccessToKlagebehandling(klagebehandlingId)
 
         return meldingMapper.toCreatedView(
             meldingService.addMelding(
@@ -66,6 +70,8 @@ class KlagebehandlingMeldingController(
         val innloggetIdent = innloggetSaksbehandlerRepository.getInnloggetIdent()
         logKlagebehandlingMethodDetails("getMeldinger", innloggetIdent, klagebehandlingId, logger)
 
+        validateAccessToKlagebehandling(klagebehandlingId)
+
         return meldingMapper.toMeldingerView(meldingService.getMeldingerForKlagebehandling(klagebehandlingId))
     }
 
@@ -80,6 +86,8 @@ class KlagebehandlingMeldingController(
     ) {
         val innloggetIdent = innloggetSaksbehandlerRepository.getInnloggetIdent()
         logKlagebehandlingMethodDetails("deleteMelding", innloggetIdent, klagebehandlingId, logger)
+
+        validateAccessToKlagebehandling(klagebehandlingId)
 
         meldingService.deleteMelding(
             klagebehandlingId,
@@ -101,6 +109,8 @@ class KlagebehandlingMeldingController(
         val innloggetIdent = innloggetSaksbehandlerRepository.getInnloggetIdent()
         logKlagebehandlingMethodDetails("modifyMelding", innloggetIdent, klagebehandlingId, logger)
 
+        validateAccessToKlagebehandling(klagebehandlingId)
+
         return meldingMapper.toModifiedView(
             meldingService.modifyMelding(
                 klagebehandlingId,
@@ -109,6 +119,10 @@ class KlagebehandlingMeldingController(
                 input.text
             )
         )
+    }
+
+    private fun validateAccessToKlagebehandling(klagebehandlingId: UUID) {
+        klagebehandlingService.getKlagebehandling(klagebehandlingId)
     }
 
 }
